refactor: flatten project structure, drop classes/ umbrella
- classes/cc/ → sources/{csv,api}/ (split: csv parsers vs API sources)
- classes/db/generics/ → db/ (flat: base.py + finance_base.py)
- classes/config.py → config.py (project root, fix .env path)
- merge file_based/online base classes → sources/base.py
