+++ sources/base.py
@@ -3,6 +3,8 @@ from abc import ABC, abstractmethod
 
 import polars as pl
 
+from config import Config
+
 
 class FileBasedCardStatement(ABC):
     type: str
@@ -23,6 +25,24 @@ class FileBasedCardStatement(ABC):
         if not os.path.exists(self.file_path):
             raise FileNotFoundError(f"File {self.file_path} not found.")
         return True
+
+    @abstractmethod
+    def load_data(self) -> None:
+        pass
+
+    def get_df(self) -> pl.DataFrame:
+        return self.df
+
+
+class OnlineCardStatement(ABC):
+    type: str
+    df: pl.DataFrame
+    config: Config
+
+    def __init__(self, type: str):
+        self.type = type
+        self.config = Config(debug=True)
+        self.load_data()
 
     @abstractmethod
     def load_data(self) -> None:
