+++ sources/registry.py
@@ -10,79 +10,79 @@ from importlib import import_module
 # Card type registry with metadata
 CARD_TYPES = {
     "amex": {
-        "module": "classes.cc.amex",
+        "module": "sources.csv.amex",
         "class_name": "AmexStatement",
         "requires_file": True,
         "description": "American Express",
     },
     "amex_annual": {
-        "module": "classes.cc.amex",
+        "module": "sources.csv.amex",
         "class_name": "AmexAnnualStatement",
         "requires_file": True,
         "description": "American Express Annual",
     },
     "rogers": {
-        "module": "classes.cc.rogers",
+        "module": "sources.csv.rogers",
         "class_name": "RogersStatement",
         "requires_file": True,
         "description": "Rogers Mastercard",
     },
     "simplii_visa": {
-        "module": "classes.cc.simplii_visa",
+        "module": "sources.csv.simplii_visa",
         "class_name": "SimpliiVisaStatement",
         "requires_file": True,
         "description": "Simplii Visa",
     },
     "simplii_debit": {
-        "module": "classes.cc.simplii_debit",
+        "module": "sources.csv.simplii_debit",
         "class_name": "SimpliiDebitStatement",
         "requires_file": True,
         "description": "Simplii Debit",
     },
     "bmo": {
-        "module": "classes.cc.bmo",
+        "module": "sources.csv.bmo",
         "class_name": "BMOStatement",
         "requires_file": True,
         "description": "BMO",
     },
     "canadian_tire": {
-        "module": "classes.cc.canadian_tire",
+        "module": "sources.csv.canadian_tire",
         "class_name": "CanadianTireStatement",
         "requires_file": True,
         "description": "Canadian Tire",
     },
     "cibc_mc": {
-        "module": "classes.cc.cibc_mc",
+        "module": "sources.csv.cibc_mc",
         "class_name": "CibcMcStatement",
         "requires_file": True,
         "description": "CIBC Mastercard",
     },
     "rbc_cc": {
-        "module": "classes.cc.rbc_cc",
+        "module": "sources.csv.rbc_cc",
         "class_name": "RbcCcStatement",
         "requires_file": True,
         "description": "RBC Credit Card",
     },
     "td_debit": {
-        "module": "classes.cc.td_debit",
+        "module": "sources.csv.td_debit",
         "class_name": "TdDebitStatement",
         "requires_file": True,
         "description": "TD Debit",
     },
     "td_visa": {
-        "module": "classes.cc.td_visa",
+        "module": "sources.csv.td_visa",
         "class_name": "TdVisaStatement",
         "requires_file": True,
         "description": "TD Visa",
     },
     "ws_debit": {
-        "module": "classes.cc.wealthsimple_debit",
+        "module": "sources.api.wealthsimple_debit",
         "class_name": "WealthsimpleDebitStatement",
         "requires_file": False,
         "description": "Wealthsimple Debit",
     },
     "ws_credit": {
-        "module": "classes.cc.wealthsimple_credit",
+        "module": "sources.api.wealthsimple_credit",
         "class_name": "WealthsimpleCreditStatement",
         "requires_file": False,
         "description": "Wealthsimple Credit",
